move finetune client to its own file
