+++ cohere/__init__.py
@@ -1,4 +1,5 @@
 from cohere.client import Client
+from cohere.finetune_client import FinetuneClient
 from cohere.client_async import AsyncClient
 from cohere.error import CohereAPIError, CohereConnectionError, CohereError
