+++ cohere/responses/finetune.py
@@ -20,13 +20,13 @@ FINETUNE_STATUS = Literal[
     "QUEUED",
 ]
 
-_INTERNAL_FINETUNE_TYPE = Literal[
+INTERNAL_FINETUNE_TYPE = Literal[
     "GENERATIVE",
     "CONTRASTIVE",
     "CLASSIFICATION",
 ]
 FINETUNE_TYPE = Literal["GENERATIVE", "EMBED", "CLASSIFY"]
-FINETUNE_PRODUCT_MAPPING: dict[FINETUNE_TYPE, _INTERNAL_FINETUNE_TYPE] = {
+FINETUNE_PRODUCT_MAPPING: dict[FINETUNE_TYPE, INTERNAL_FINETUNE_TYPE] = {
     "GENERATIVE": "GENERATIVE",
     "EMBED": "CONTRASTIVE",
     "CLASSIFY": "CLASSIFICATION",
